feat: remove Trackers
Reasons for this are multiple. I was not happy with how the syntax
turned out, it's hard to keep track of that much data in a shape that's
usable in a Markdown file (e.g. a daily tracker after a year is 365
lines). If anywone were to complain about this removal, then I would
consider rethinking and reimplementing it. As of today, even I, don't
want to use the feature.